refactor, wip tests
