+++ app/src/main/java/com/raygun/raygun4android/sample/fragments/Fragment1.kt
@@ -1,0 +1,21 @@
+package com.raygun.raygun4android.sample.fragments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.raygun.raygun4android.sample.R
+
+/**
+ * A simple empty [Fragment].
+ */
+class Fragment1 : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_1, container, false)
+    }
+}
