+++ app/src/main/res/layout/activity_navigation.xml
@@ -5,13 +5,13 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NavigationActivity">
+    tools:context=".fragments.NavigationActivity">
 
     <!-- res/layout/example_activity.xml -->
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:name="com.raygun.raygun4android.sample.Fragment1" />
+        android:name="com.raygun.raygun4android.sample.fragments.Fragment1" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
